pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[tool.flit.module]
name = "sphinx_localtimezone" # FIXME

[project]
name = "sphinx-localtimezone" # FIXME
authors = [{name = "Richard Darst"}]  # FIXME
readme = "README.md"
license = {file = "LICENSE"}
# https://pypi.org/classifiers/
classifiers = [
    "License :: OSI Approved :: MIT License",
    ]
#keywords = ["sphinx-extension"]
requires-python = ">=3.6"
dynamic = ["version", "description"]
dependencies = [
    # FIXME
    "Sphinx",
    "python-dateutil",
]

[project.optional-dependencies]
test = [
    # FIXME
    "pytest",
]


#[project.urls]
#Repository = "https://example.com/"
#Documentation = "https://example.com/"


# https://flit.pypa.io/en/latest/upload.html
# flit build
# You need to configure a .pypirc file for test upload, or use environment variables:
# https://flit.pypa.io/en/latest/upload.html#using-pypirc
# flit publish --repository testpypi
# or:   FLIT_INDEX_URL=https://test.pypi.org/legacy/ FLIT_USERNAME=xxx and FLIT_PASSWORD=xxx flit publish
# flit publish
